runtime: Save context when entering exception
When we call the function to enter the excpeption stack, we should
preserve the current context to resume execution at this point. This
is useful for the functions that are called directly from kernel
threads. The function now returns 1 when called. The context contains
an execution point in this function and returns 0 when restored.