runtime: Search for content of data structure
Until now we have only been able to search for the content of data,
i.e., search for a value if integers are stored in the list or for a
pointer value if pointers to data structures were stored in the list.

With optimsoc_list_find_content_header it is now possible to search
for the element that points to this value. This means if you store
uint32_t* values in your list, you can search for a uint32_t value and
get the pointer to this value.

More importantly you can use it to find indexed data structures, such
as

struct yourstruct {
       uint32_t index;
       /* other elements in struct */
}

As the function dereferences the pointer as an integer, it will here
compare to the first element in your struct, which is very useful to
use sparse lists.